Refactor admin contract into different files
